requirements.txt
google-generativeai>=0.9.0
python-dotenv
pillow